+++ flexiapi/resources/views/api/documentation_markdown.blade.php
@@ -428,12 +428,14 @@ Return the updated account.
 
 ## Accounts devices
 
-### `GET /accounts/me/devices`
+### `GET /accounts/{id/me}/devices`
+<span class="badge badge-warning">Admin</span>
 <span class="badge badge-info">User</span>
 
 Return the user registered devices.
 
-### `DELETE /accounts/me/devices/{uuid}`
+### `DELETE /accounts/{id/me}/devices/{uuid}`
+<span class="badge badge-warning">Admin</span>
 <span class="badge badge-info">User</span>
 
 Remove one of the user registered devices.
@@ -450,9 +452,10 @@ Return the user contacts.
 
 Return a user contact.
 
-## Account vCards storage
+## vCards storage
 
-### `POST /accounts/me/vcards-storage`
+### `POST /accounts/{id/me}/vcards-storage`
+<span class="badge badge-warning">Admin</span>
 <span class="badge badge-info">User</span>
 
 Store a vCard.
@@ -461,7 +464,8 @@ JSON parameters:
 
 * `vcard`, mandatory, a valid vCard having a mandatory `UID` parameter that is uniquelly identifying it. This `UID` parameter will then be used to manipulate the vcard through the following endpoints as `uuid`.
 
-### `PUT /accounts/me/vcards-storage/{uuid}`
+### `PUT /accounts/{id/me}/vcards-storage/{uuid}`
+<span class="badge badge-warning">Admin</span>
 <span class="badge badge-info">User</span>
 
 Update a vCard.
@@ -470,17 +474,20 @@ JSON parameters:
 
 * `vcard`, mandatory, a valid vCard having a mandatory `UID` parameter that is uniquelly identifying it and is the same as the `uuid` parameter.
 
-### `GET /accounts/me/vcards-storage`
+### `GET /accounts/{id/me}/vcards-storage`
+<span class="badge badge-warning">Admin</span>
 <span class="badge badge-info">User</span>
 
 Return the list of stored vCards
 
-### `GET /accounts/me/vcards-storage/{uuid}`
+### `GET /accounts/{id/me}/vcards-storage/{uuid}`
+<span class="badge badge-warning">Admin</span>
 <span class="badge badge-info">User</span>
 
 Return a stored vCard
 
-### `DELETE /accounts/me/vcards-storage/{uuid}`
+### `DELETE /accounts/{id/me}/vcards-storage/{uuid}`
+<span class="badge badge-warning">Admin</span>
 <span class="badge badge-info">User</span>
 
 Delete a stored vCard